+++ src/Cake.Slack/Chat/SlackChatProvider.cs
@@ -1,37 +1,32 @@
-﻿using Cake.Core;
-using Cake.Core.Annotations;
+﻿using System;
+using Cake.Core;
 
-// ReSharper disable UnusedMember.Global
-// ReSharper disable MemberCanBePrivate.Global
 namespace Cake.Slack.Chat
 {
     /// <summary>
-    /// Contains Slack Chat Extensions
+    /// Contains SlackProvider Chat functionality.
     /// </summary>
-    [CakeAliasCategory("Slack.Chat")]
-    public static class SlackChatExtensions
+    public sealed class SlackChatProvider
     {
+        private readonly ICakeContext _context;
+        
         /// <summary>
         /// Post message to Slack Channel
         /// </summary>
-        /// <param name="context">Cake context</param>
-        /// <param name="token">Slack auth token</param>
+        /// <param name="token">SlackProvider auth token</param>
         /// <param name="channel">Channel to send message to. Can be a public channel, private group or IM channel. Can be an encoded ID, or a name.</param>
         /// <param name="text">Text of the message to send. For message formatting see https://api.slack.com/docs/formatting</param>
         /// <returns>Returns success/error/timestamp <see cref="SlackChatMessageResult"/></returns>
-        [CakeMethodAlias]
-        public static SlackChatMessageResult SlackChatPostMessage(
-            this ICakeContext context,
+        public SlackChatMessageResult PostMessage(
             string token,
             string channel,
             string text
             )
         {
-            return context.SlackChatPostMessage(
-                token,
+            return _context.PostMessage(
                 channel,
                 text,
-                new SlackChatMessageSettings()
+                new SlackChatMessageSettings {Token = token}
                 );
         }
 
@@ -39,27 +34,35 @@ namespace Cake.Slack.Chat
         /// <summary>
         /// Post message to Slack Channel
         /// </summary>
-        /// <param name="context">Cake context</param>
-        /// <param name="token">Slack auth token</param>
         /// <param name="channel">Channel to send message to. Can be a public channel, private group or IM channel. Can be an encoded ID, or a name.</param>
         /// <param name="text">Text of the message to send. For message formatting see https://api.slack.com/docs/formatting</param>
-        /// <param name="messageSettings">Lets you override default settings like UserName, IconUrl or if it should ThrowOnFail</param>
         /// <returns>Returns success/error/timestamp <see cref="SlackChatMessageResult"/></returns>
-        [CakeMethodAlias]
-        public static SlackChatMessageResult SlackChatPostMessage(
-            this ICakeContext context,
-            string token,
+        /// <param name="messageSettings">Lets you override default settings like UserName, IconUrl or if it should ThrowOnFail</param>
+        public SlackChatMessageResult PostMessage(
             string channel,
             string text,
             SlackChatMessageSettings messageSettings
             )
         {
-            return context.PostMessage(
-                token,
+            if (messageSettings == null)
+            {
+                throw new ArgumentNullException("messageSettings");
+            }
+
+            return _context.PostMessage(
                 channel,
                 text,
                 messageSettings
                 );
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="SlackChatProvider"/> class.
+        /// </summary>
+        /// <param name="context">The context.</param>
+        public SlackChatProvider(ICakeContext context)
+        {
+            _context = context;
         }
     }
 }