+++ ReleaseNotes.md
@@ -1,2 +1,5 @@
+### New in 0.0.2 (Released 2015/06/27)
+* Added support to post message using incoming Slack web hook
+* Refactored to property alias (method alias kept for backward compatibility)
 ### New in 0.0.1 (Released 2015/01/13)
 * First release of Cake.Slack